Move DimExpr Simplify and Substitute to dim_expr_util
