+++ paddle/cinn/common/broadcast_tree.cc
@@ -17,8 +17,7 @@
 #include <optional>
 #include <unordered_map>
 
-#include "paddle/pir/include/dialect/shape/utils/dim_expr_simplify.h"
-#include "paddle/pir/include/dialect/shape/utils/dim_expr_substitute.h"
+#include "paddle/pir/include/dialect/shape/utils/dim_expr_util.h"
 
 namespace cinn::common {
