+++ paddle/pir/include/dialect/shape/utils/shape_or_data_expr.h
@@ -15,7 +15,7 @@
 #pragma once
 
 #include "paddle/pir/include/dialect/shape/utils/dim_expr.h"
-#include "paddle/pir/include/dialect/shape/utils/dim_expr_simplify.h"
+#include "paddle/pir/include/dialect/shape/utils/dim_expr_util.h"
 
 namespace symbol {
