+++ paddle/pir/src/dialect/shape/utils/dim_expr_builder.cc
@@ -14,7 +14,7 @@
 
 #include "paddle/pir/include/dialect/shape/utils/dim_expr_builder.h"
 #include "paddle/common/enforce.h"
-#include "paddle/pir/include/dialect/shape/utils/dim_expr_simplify.h"
+#include "paddle/pir/include/dialect/shape/utils/dim_expr_util.h"
 
 namespace symbol {
