+++ test/cpp/pir/shape_dialect/simplify_dim_expr_test.cc
@@ -14,7 +14,7 @@
 
 #include <atomic>
 #include "gtest/gtest.h"
-#include "paddle/pir/include/dialect/shape/utils/dim_expr_simplify.h"
+#include "paddle/pir/include/dialect/shape/utils/dim_expr_util.h"
 
 namespace symbol::test {
